Fixed Inventory script
Inventory was should 5/5 keys for both Level_1 and Level_2 but i created a method using read only properties to check the number of total keys needed to be collected in each scene that was loaded before exiting scenes.

Need to do more fixes before submission, time was blocked due to no internet for weeks.